+++ NEWS.md
@@ -17,8 +17,9 @@ Version 1.0.1b1 - dev
   fix minimum axis-ratio for the ELLIPSE entity, added upperbound tolerance to axis-ratio 
   validator to take floating point imprecision into account
 - BUGFIX: [#810](https://github.com/mozman/ezdxf/issues/810)
-  fix function `ezdxf.render.forms.cylinder_2p()` for cylinder axis parallel to the 
-  z-axis 
+  fix function `ezdxf.render.forms.cylinder_2p()` for cylinder axis parallel to z-axis 
+- BUGFIX: [#811](https://github.com/mozman/ezdxf/issues/811)
+  fix function `ezdxf.render.forms.cone_2p()` for cone axis parallel to z-axis 
 
 
 Version 1.0.0 - 2022-12-09